requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
google-generativeai>=0.7.0
pillow
streamlit
python-dotenv
requests
